requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
fonttools==4.39.3
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numba==0.57.0
numpy==1.24.3
omegaconf==2.3.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip==23.0.1
platformdirs==3.4.0
prompt-toolkit==3.0.38
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==305.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
tslearn==0.5.3.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
wheel==0.38.4
zipp==3.15.0
